+++ core/src/main/kotlin/JavaFeatures.kt
@@ -10,7 +10,7 @@ import edu.illinois.cs.cs125.jeed.core.antlr.JavaParser.StatementContext
 import edu.illinois.cs.cs125.jeed.core.antlr.JavaParserBaseListener
 import org.antlr.v4.runtime.ParserRuleContext
 import org.antlr.v4.runtime.Token
-import org.antlr.v4.runtime.tree.ParseTreeWalker
+import org.antlr.v4.runtime.tree.IterativeParseTreeWalker
 import org.antlr.v4.runtime.tree.TerminalNode
 
 internal val seenJavaFeatures = mutableSetOf<FeatureName>()
@@ -850,6 +850,6 @@ class JavaFeatureListener(val source: Source, entry: Map.Entry<String, String>) 
     init {
         val parsedSource = source.getParsed(filename)
         // println(parsedSource.tree.format(parsedSource.parser))
-        ParseTreeWalker.DEFAULT.walk(this, parsedSource.tree)
+        IterativeParseTreeWalker().walk(this, parsedSource.tree)
     }
 }