+++ core/src/main/kotlin/KotlinMutation.kt
@@ -6,7 +6,7 @@ import edu.illinois.cs.cs125.jeed.core.antlr.KotlinParser.*
 import edu.illinois.cs.cs125.jeed.core.antlr.KotlinParserBaseListener
 import org.antlr.v4.runtime.ParserRuleContext
 import org.antlr.v4.runtime.RuleContext
-import org.antlr.v4.runtime.tree.ParseTreeWalker
+import org.antlr.v4.runtime.tree.IterativeParseTreeWalker
 import org.antlr.v4.runtime.tree.TerminalNode
 import org.jetbrains.kotlin.backend.common.pop
 
@@ -729,7 +729,7 @@ class KotlinMutationListener(private val parsedSource: Source.ParsedSource) : Ko
 
     init {
         // println(parsedSource.tree.format(parsedSource.parser))
-        ParseTreeWalker.DEFAULT.walk(this, parsedSource.tree)
+        IterativeParseTreeWalker().walk(this, parsedSource.tree)
         check(loopDepth == 0)
         check(loopBlockDepths.isEmpty())
     }